requirements.txt
llama-index==0.11.15
llama-index-vector-stores-postgres==0.2.5
llama-index-embeddings-ollama
docling==2.53.0
pypdf==4.3.1
python-dotenv==1.0.1
tqdm==4.66.4
fastapi
uvicorn
requests
llama-index-llms-ollama
psycopg2-binary
